// src/generated/github.ts
export enum SubscriptionState {
  /** The User is never notified. */
  IGNORED = 'IGNORED',
  /** The User is notified of all conversations. */
  SUBSCRIBED = 'SUBSCRIBED',
  /** The User is only notified when participating or @mentioned. */
  UNSUBSCRIBED = 'UNSUBSCRIBED'
}